+++ CHANGELOG.md
@@ -63,6 +63,7 @@
 - Retry 500 responses from objectstore. ([#6162](https://github.com/getsentry/relay/pull/6162))
 - Remove the `metrics_extracted` and `spans_extracted` flags from the transactions processing pipeline. ([#6190](https://github.com/getsentry/relay/pull/6190), [#6200](https://github.com/getsentry/relay/pull/6200))
 - Convert TUS uploads to Objectstore multipart uploads. ([#6172](https://github.com/getsentry/relay/pull/6172))
+- Stop accepting the deprecated Expect-CT, HPKP, and Expect-Staple security reports. ([#6230](https://github.com/getsentry/relay/pull/6230))
 
 ## 26.6.0